// src/types/partials/NutritionOrderEnteralFormula.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_classes so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TQuantity } from '../partials/Quantity';
import { TNutritionOrderAdministration } from '../partials/NutritionOrderAdministration';

export type TNutritionOrderEnteralFormula = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    baseFormulaType?: TCodeableConcept;
    baseFormulaProductName?: String;
    additiveType?: TCodeableConcept;
    additiveProductName?: String;
    caloricDensity?: TQuantity;
    routeofAdministration?: TCodeableConcept;
    administration?: TNutritionOrderAdministration[];
    maxVolumeToDeliver?: TQuantity;
    administrationInstruction?: String;
};